Application: Separate base code into BaseApplication
